GH-661 join addon-manager and gw-controller binaries
