+++ bundle/manifests/mgc-add-on-manager_v1_serviceaccount.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  creationTimestamp: null
+  labels:
+    app.kubernetes.io/component: rbac
+    app.kubernetes.io/created-by: kuadrant
+    app.kubernetes.io/instance: add-on-manager
+    app.kubernetes.io/managed-by: kustomize
+    app.kubernetes.io/name: serviceaccount
+    app.kubernetes.io/part-of: multicluster-gateway-controller
+  name: mgc-add-on-manager
