+++ bundle/manifests/mgc-kuadrant-addon_rbac.authorization.k8s.io_v1_clusterrolebinding.yaml
@@ -1,0 +1,13 @@
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  creationTimestamp: null
+  name: mgc-kuadrant-addon
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: mgc-kuadrant-addon
+subjects:
+- kind: ServiceAccount
+  name: mgc-add-on-manager
+  namespace: multicluster-gateway-controller-system
