+++ cmd/gateway_controller/addon-manager/manifests/cluster-role.yaml
