+++ cmd/gateway_controller/addon-manager/manifests/kuadrant-namespace.yaml
