+++ cmd/gateway_controller/addon-manager/manifests/operator-group.yaml
