+++ cmd/gateway_controller/addon-manager/manifests/subscription.yaml
