--- config/add-on-manager/cluster-management-addon.yaml
@@ -1,10 +1,0 @@
-apiVersion: addon.open-cluster-management.io/v1alpha1
-kind: ClusterManagementAddOn
-metadata:
- name: kuadrant-addon
-spec:
- addOnMeta:
-   displayName: kuadrant Addon
-   description: "kuadrant operator"
-
-
