--- config/add-on-manager/manager.yaml
@@ -1,49 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: add-on-manager
-  namespace: system
-  labels:
-    control-plane: kuadrant-add-on-manager
-spec:
-  selector:
-    matchLabels:
-      control-plane: kuadrant-add-on-manager
-  replicas: 1
-  template:
-    metadata:
-      annotations:
-        kubectl.kubernetes.io/default-container: controller
-      labels:
-        control-plane: kuadrant-add-on-manager
-    spec:
-      securityContext:
-        runAsNonRoot: true
-      containers:
-      - command:
-        - /add-on-manager
-        args:
-        - --leader-elect
-        image: addon-manager:latest
-        imagePullPolicy: Always
-        envFrom:
-          - configMapRef:
-              name: controller-config
-              optional: true        
-        name: controller
-        securityContext:
-          allowPrivilegeEscalation: false
-          capabilities:
-            drop:
-              - "ALL"
-        #TODO add health and readiness probes 
-        # More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
-        resources:
-          limits:
-            cpu: 500m
-            memory: 128Mi
-          requests:
-            cpu: 10m
-            memory: 64Mi
-      serviceAccountName: add-on-manager
-      terminationGracePeriodSeconds: 10
