+++ config/default/kustomization.yaml
@@ -11,7 +11,6 @@ namePrefix: mgc-
 resources:
 - ../rbac
 - ../manager
-- ../add-on-manager
 
 patches:
 - path: manager_metrics_patch.yaml
